Write GAIN before ADSR when setting up instrument
Due to being called to update writing DSP registers, GAIN is written after the
ADSR registers due to the order they are read and written to. This is not
desirable if the GAIN is being changed and the ADSR is being set to GAIN mode at
the same time. Thus, the read and write order is changed here.

This commit closes #25.