Reallocate SPECIALWAVE.brr in #AMM to first sample slot
Turns out there was a misunderstanding about the #AMM sample group: its only
purpose is to provide padding so that sample IDs $20 and up are allocated to the
custom samples used in an AddMusicM song. Two new sample groups were going to be
created via overwriting sample ID $09, but there is a conflict with the bonus
end song, a default global song at that, meaning that it would not be feasible
to have this done by default.

Thus, SPECIALWAVE.brr has been reallocated from sample ID $1D to sample ID $14,
with $1D being the starting expected sample ID plus $09, which was the source of
the original misinterpretation... as a zero was used.