Force !SFXEchoChannels to zero when initializing a song
This looks like it was the intended method. A possible bug was accidentally
introduced when the slowdown fix was implemented where if a tick should
immediately fire after a long time since the timer was last read, then the
channels set up for using SFX would end up non-zero.

This commit closes #39.