Remap grinder SFX to CPUIO1-accurate values and add 1DF9/1DFC switch
SFX ID $1A in $1DF9 was always a duplicate even in vanilla SMW, and it was
redefined in AddMusicK to this duplicate value initially. The new CPUIO1 SFX
replicas have priorities attached to them, whereas this one doesn't: thus, the
IDs used have been moved, and the ability to switch the allocation of the
grinder SFX to 1DFC has also been added.