Don't clear pitch envelope variables when initializing SFX
In CPUIO0 and CPUIO3's case, the wrong pitch envelope memory location was being
cleared, causing an edge case when the envelope would still run AND causing the
entire envelope to not be restored after the SFX was finished playing.

CPUIO1 had the same problem as CPUIO0 and CPUIO3 from a restoration standpoint,
and thus its clear loop, which was initially shrunk to get rid of clearing out
arpeggio variables, is now completely gone.