Initialize instrument table properly when restoring from CPUIO1 SFX
The instrument table was not being initialized properly after a CPUIO1 SFX was
done playing. The jump location was the cause, so it was corrected. In addition,
some code consolidation was done, as the grinder SFX (CPUIO1 command ID $04) was
doing even less to restore the SFX: namely, it didn't even account for an
AMK-specific flag!

This commit closes #49.