Implement speed-up effect for P-Switch SFX
The P-Switch SFX now has a speed-up effect like the actual song, but it's a bit
more limited due to there being no tempo slide command. Instead, the note
lengths are manually set on a per-loop basis. Thanks to everything being nice
and divisible by the shortest note length, this isn't too much of a hassle.
Note that the programming accounts for channel interruptions and the order for
which the SFX is processed in channel-wise. This is because unlike the music,
the SFX are normally independent of each other in execution, meaning that some
extra effort needs to be done if one wants to modify note lengths and keep them
in sync.