Replace arch spc700-raw references for future asar versions
arch spc700-raw is going to get removed in a future asar version.
Thus, they have been replaced with arch spc700 along with a norom mapping mode
specifier.

Due to norom only working with the 65816 architecture at this time, the end
result is a bit on the ugly side as far as what is expected to appear in the
code. arch 65816 will be removed when a version of asar with this bug removed
appears on SMWCentral.

This commit closes #66.