Preserve noise clock during echo VCMDs
Small piece of fine print: when setting up the EDL, if the entire sound has to
be muted, the noise clock is allowed to temporarily become zero. It is restored
after the setup is done, though.
This commit closes #16, closes #76 and closes #77. That's a three-in-one result
due to the intermingling of these three going together due to the code involved.