Terminate SFX sequence and don't disable SFX flag when CPUIO1 SFX triggers
CPUIO1 SFX conveniently has a timer present to wait for the SFX to start up,
meaning the sequence can be terminated without too many interruptions
occurring. However, this requires an extra check being coded to detect a CPUIO1
SFX on standby, otherwise the CPUIO1 SFX is in danger of being overwritten by
the music.

This commit closes #79.