Allow non-SFX APU commands
CPUIO0 and CPUIO1 have commands that don't require SFX to play. A special
adaptation was made for the tempo hike effect. The pause command is currently
not included due to embedding its ASM within the SFX instances: this can be
corrected by moving them to main.asm, but it won't be done in this commit for
now.