Correct possible channel overflow bug when setting volume from noise setting
The X register can end up reading the track pointer for the repeat section of
channel 0 ($01E0) by mistake for one iteration only if it is already on the
highest channel when the priorities are already equal in the first place. This
check is added on to avoid that particular edge case.