Mark all user-defined samples in music as important
The user is expecting all user-defined samples to actually be added in to the
final result. Most commonly they're manually taken from the #optimized sample
group without selecting all samples. This is generally happening for ROM
compilation (though with post-filtering afterwards), but it is not happening for
SPC compilation.

This commit closes #121.