Import important markers from samples taken from sample groups
The user can define a sample taken straight out of a sample group without taking
the entire sample pack in one go. However, this tends to result in inconsistency
between SPC and ROM compilation as far as what samples are included because of
an external factor (specifically, a song using samples from a sample group
before a song using individually picked samples chooses them). Thus, the default
behavior is now to import the important flag from the group if the sample is
part of a sample group.

This commit closes #123.