Add Visualizations path
This path is never pre-created in the original AddmusicK releases, and there is
a (currently) undocumented feature that uses this folder, debuting in AMK 1.0.2.
Said feature is probably one of the last things that was ever implemented by the
original author, as 1.0.2 is the last known version that the original author
worked on.