Use a branching distance gate embedded in code for runningArp
RAM in the $0000-$0400 is getting increasingly scarce, so in some cases it's
more practical to use what I call "gates". They're already used in another
branch for the special pulse wave, but this kind of gate changes the branching
distance depending on whether or not the next note was generated by an arpeggio
note. This causes runningArp to no longer take a byte in the RAM section, since
it's just one bit being used out of eight in the first place.