Added the ability to have jump and grinder SFX default to 1DFA
Previously AddMusicK had completely cut out the 1DFA versions of these SFX.
Because various improvements had been made to make the code play nice with the
SFX sequences and music data, they have been restored by acting as the default
case. Two new defines were added to keep support for 1DF9 for the jump and
grinder SFX.

Note that this situation only happens when 1DF9 and 1DFC are both not used for
either SFX: otherwise, the situation is the same as before, and the defaults
have not been modified: instead, this just simply opens up the possibility of
using vanilla SMW values.