Set priority to match AMK 1.0.8's priority system
Turns out the priority system, which was designed to replicate vanilla SMW's,
ran into a few complications caused by using only two channels instead of three.
Therefore, the priorities have been commented out of most of the ones that were
added on to in the first place, and 1DFA's priorities have been changed to
utilize a user definition, defaulting to zero but providing a vanilla value.

In the future, a command line option, along with making sure defines carry
either from asar to SFX instances or across SFX instances, will ensure that
these can be restored as needed. For now, for the SFX affected, simply
uncommenting them will be good enough.