Change memory locations used by $F7 VCMD for indirects
This VCMD, taken from AddmusicM, was never updated to account for the fact that
direct page $20 is now reserved for SFX data. Therefore, this has been
corrected. However, it's still disabled by default, since most of the memory
locations are de facto not stable at this time since the builds are subject to
constant modification.