Don't include SFX data when !noSFX is set to !true
All SFX data must not be included when !noSFX is turned on. Thus, the SFXTable
labels have been targeted, which performs this task correctly. However...
The C++ side still says the total filesize of the main program + all sound
effects. This define needs to be detected on the C++ side to avoid confusing the
user.