Correct a slight error on the start position of the sound engine code
The sound engine code's starting position is slightly off due to using the
same position that was used that caused #54. This isn't accurate to the actual
start of the sound engine code, and thus falsely reports some extra bytes as
being occupied by variable data, when in reality there are already opcodes
there.