Replace AM4Batch and AMMBatch with HertzDevil's updates
These two are also using Windows-specific code, and therefore are being
replaced. This leaves AMKGUI as the last one standing, and that's because it has
Windows-specific GUI code, meaning it won't be replaced unlike the others.
These files have been adapted for the latest version of Visual Studio as well.

Original sources:
https://github.com/HertzDevil/AddmusicK/tree/master/AM4Batch
https://github.com/HertzDevil/AddmusicK/tree/master/AMMBatch

There are several source commits (but only the AM4Batch and AMMBatch files are
used):
https://github.com/HertzDevil/AddmusicK/commits/master/AM4Batch
https://github.com/HertzDevil/AddmusicK/commits/master/AMMBatch
https://github.com/HertzDevil/AddmusicK/commit/af66dff5f943685db71f743a6acb93107a27cd3d
https://github.com/HertzDevil/AddmusicK/commit/4e0937c14f5749e71ddda588989106e31624d63c

This commit closes #163.