Add noSFX detection and handling on C++-side
Currently noSFX is handled exclusively through the asar side of things, and not
anywhere in the C++ code. Some of the terminology might be inaccurate if we
don't modify the C++ side on the final results... and SPC dumps of SFX would be
completely non-functional, meaning this would require that a stopgap be
implemented.