Make memory location utilizations easier to track in SyncInc
Sometimes things get way tricker than they're supposed to be when it comes to
analyzing what memory locations are supposed to be used where, simply because
the way it is specified in the assembly is a bit non-intuitive. Although the
final compiled result can't be helped in this fashion, the raw assembly code to
compile, on the other hand, can, so that the user can more easily search for
what memory locations are used where.

SyncInc used a setp/clrp pair, so that has been accounted for here.