Optimize writing to pitch and volume DSP registers
The calculation routines have been set to terminate immediately if any SFX is
playing on the channel, rather than do all that work and merely bypass writing
to the DSP registers. This is because they're only touching scratch RAM at this
point: thus, they need not apply if they're not going to get sent to the DSP
registers.