Auto-fetch first VCMD parameter byte, like modern N-SPC/Kankichi
Fetching the first byte after the initial command ID, using the length table to
verify that one does exist, allows for the removal of quite a few GetCommandData
calls at the beginning, since it's now done beforehand. There were a couple of
minor reprogramming cases to account for this modification, but nothing too
major. As a side effect of doing this, $EB and $EC actually lose their
optimization from earlier because A ends up non-zero due to fetching a byte
first.