Unconditionally restore the music's noise frequency when SFX is done
There is a case where no channels will have their volumes zeroed out because
they are overwritten by the SFX that also uses the noise. In these cases, the
music noise frequency must be unconditionally restored, or else the channels
will not be able to be restored in this fashion.

This commit closes #183.