Have $FF actually play the last note endlessly
Turns out other that other than having to rewind a byte anyways, all of the
parameters that this VCMD would theoretically need (except for the note value,
which would simply have the save VxPITCH value as before) in order to endlessly
play the last note, length and all, are actually already stored away and are
ready to use on the fly. Thus, instead of fetching the bytes for the last note
and having to scan backwards for the last valid note ID (that would involve a
lot of ID checking going backwards to make sure we don't mistake it for a VCMD),
we instead simply take the saved data from the last note processed and use that
instead.

This commit closes #75.