Modify !NCKValue to prevent FLG registers from being overwritten
The noise SFX frequency conflict resolution code, specifically the one that
restores the music noise frequency, is causing problems with the pause SFX by
overwriting the mute flag on the FLG register. Therefore, we have to actually
set the flag in question on the mirror as well.

For the pause SFX, this actually makes directly setting the FLG register
beforehand redunant, since it happens anyways at the end. For the both unpause
SFX variants, setting the FLG DSP register is not optional, and thus we call
ModifyNoise right then and there to get that set up.

This commit closes #194.