Don't force percussion notes on Addmusic405 and AddmusicM ports on Ch6&Ch7
This wasn't the behavior of the original programs, and thus counts as a
conversion bug. The correct behavior is to only do so on the first note:
AddmusicK added some code that always outputs percussion notes on #6 and #7, but
this conflicts with the original behavior on AM405 and AMM. So the code now only
applies to when AddmusicK itself is the target program.

This commit closes #198.