Fix fatal bug with lowest non-zero channel number with Addmusic405 songs
The phrase end marker was being generated for Ch0 even if the channel was never
defined in the first place. For Addmusic405 songs, this is a fatal bug, since it
causes the sound driver to stop in its tracks as it endlessly restarts
the phrase. The scope has been limited to Addmusic405 for now: other targets
will be handled on a case-by-case basis if they are encountered, and AMK itself
will not have this applied to it for consistency's sake.

This commit closes #202.