Add arpeggio ($FB) VCMD filtering
Arpeggio takes up a good chunk of memory to implement, and until AddmusicK 1.0.9
(RC), it wasn't very functional, either. Thus, by setting !noVcmdFB to !true,
the code will be removed from the sound driver on compilation.

Please note that pass-through code has not been properly programmed in yet, so
the sound driver will crash if you try to use this VCMD with !noVcmdFB set to
true!

This commit mentions #85.