Use hard-coded transposition for built-in instruments on AddmusicM
AddmusicM overwrites the h values by default when handling transposition, which
previously was not handled correctly in AddmusicK. This is only a problem with
AddmusicM due to it implementing a new transposition command during compilation:
Addmusic405 uses a per-track transposition command that only affects the SPC700
side, and Addmusic Java notes, but never implements, a transposition command.

This commit closes #218.