Split remote code clear code into two separate type of clears
In order for the upcoming modification to work with AMK beta's handling of gain,
we need to split the remote code clearings into two types. The reason for this
is that the slot normally reserved for KON remote commands is going to be used
due to the requirement of remote gain and anticipation gain being able to run at
the same time.

This commit mentions #225.