Revamp remote code event type 5
Event type 5, which was reserved for gain conversions in the first place, now
acts like a type 2 event. However, that's not all: it contains a special auto
restore feature that normally had to be done separately. This will be critical,
as the C++ side is going to have its conversions simplified as a result of doing
this. Stay tuned, because we have one more new remote code event type that will
have this kind of special purpose...

This commit mentions #225.