Correct channel unmuting problem + optimize Yoshi Drum handling code
First off, the mute flag is now completely cleared prior to conditionally
setting any Yoshi Drum bits. This means that a section of commented out code is
included in case any external code wants to set any additional flags: they just
need to send it over to the embedded memory location in question, and then
uncomment a couple a lines of code.

Secondly, some code optimization related to Yoshi Drum handling was done. $0386
is officially losing its usage due to being turned into an opcode switch
depending on whether the Yoshi Drums are on or off: thus, the memory location is
now free.

This commit closes #230.