Updating AMK Beta ARAM Map modifications
In addition to correcting a typo, the notes have been clarified regarding remote
gain's timer value and its replacement for remote code event 2 (not 3). Plus, an
invalid memory location reference has been added on due to being present in the
AMK beta build and was coded in for the $FB VCMD, which in turn would overwrite
a memory location if used (had the high and low bytes been separated, this would
have been acceptable).