Simplify conversion of anticipation gain/remote gain events
With our new event types ready to go, it is now possible to simplify the
conversion process to not have to account for either $FA $05 or $FC having
previously been executed: they can now be independent of each other. This is
because both can now be stopped individually, rather than having to restore one
of the gain values after executing a type 0 remote code event.

This commit mentions #225. It does not close the issue because of #25, in
addition to the conversions being more optimizable. Because the original code
used direct DSP register writes, a different solution may be required to avoid
conflicts with music playback from other VCMDs.