Wait for the previous EDL to elapse before setting ESA and EDL
The method for waiting is going to get slightly modified. Namely, the echo
buffer will now be "captured" in a four byte area at the beginning of the buffer
before being moved around. This is done by waiting for at least one normal echo
buffer write cycle (that is, when it reaches the end of the buffer). Then it
will be safe to modify both the ESA and the EDL.