Don't modify !NCKValue during ModifyEchoDelay
This is a minor modification that allows !NCKValue to remain unchanged and
actually allows me to save the original state of the FLG DSP register in the
process without having to do much else. There is a clr1 opcode added as a side
effect of doing this, but this is caused by a bug in the original AddmusicK that
could have been subject to exploitation... and one which I prefer to resolve in
a different branch altogether (though after merging this one).