Disable Yoshi Drums if Mario is not on Yoshi at the start of a room
This kind of disabling was formerly implicit in Vanilla SMW since they get
disabled when going to the overworld. A new hijack was created to make disabling
the Yoshi Drums explicit when this check fails. The drums may appear for a brief
moment while the room is loading, but they do stop shortly before the level
starts.

This commit mentions #243. I'm not completely confident that it closes it, but
it's better than what it was like before.