Use standard SFX sequence data for CPUIO1 SFX
For big savings, we can have the standard SFX sequence format take over. The
original code is still present both for historical reference... and in case
someone prefers the hardcoded approach or finds a problem when using the SFX
sequence replicas. The default is now set to use these sequences due to the
amount of memory this saves.

This commit closes #249.