Update CommandDispatchTable references
The value $B4 that was originally used is accurate if you knew that the value in
question was $DA doubled then ANDed by $FF... but such a number may not
immediately be clear to anyone else looking at the code. Thus, this has been
updated to indicate that it is relative to the command ID in question. As an
added bonus, it links up quite nicely to the CommandLengthTable reference, which
uses the same value.