Output warnings for use of non-native hex commands from past Addmusics
These are not going to be pre-processed. They also give out a suggestion on
which one might be the correct one (in the case of #amk 1, it's because the user
could be in the middle of a conversion from #amm or #am4 and forgot to change
the parser flag).

This commit mentions #242.