Implement hot-patch VCMD on SPC700 side
This special VCMD is intended to replicate the behavior of past Addmusics.
Addmusic405's readahead mayhem is not implemented yet, though... and there may
be some other undiscovered differences. For AddmusicK Beta, its handling of DSP
register writes for remote and anticipation gain still needs to be adapted
because it uses voice-specific DSP register writes and doesn't use the backup
table: whether this translates into a hot patch or not is still to be
determined.

Forced patch filters will be coming later: this is just the VCMD code.

This commit mentions #222.