Don't pre-allocate bytes when setting up music
Previously, this was a relatively stable operation under the condition that the
bytes were filled in. However, some future additions could make this operation
inherently unstable, since they may not always need to be the first six bytes.
Thus, they are now generated once everything else is accounted for, which also
allows extra bytes to be inserted in between. This does mean that the pointers
need to be recalibrated, though, since they would otherwise refer to an
incorrect offset.