Auto-generate hot patch preset VCMDs for historical Addmusics
AddmusicK Beta, AddmusicM and Addmusic405 will now have a hot patch VCMD auto
generated by preset ID. These will ensure that they properly replicate the
playback of the original. Note that these will be subject to updates in the
future, but it will all be by setting and clearing bits, not by modifying the
preset IDs.

You may have noticed that AddmusicK 1.0.9 (which is the next version of
AddmusicK) is not included: that will have to be done manually. This is because
of parser version conflicts: 2 relies on AMK 1, and 3 relies on Codec's
AddmusicK 1.1.0 Beta. Thus, I can't automatically include the bugfixes I made:
instead, you'll have to manually include a $FA $7F $01 (Zero is the default, and
it represents the original finalized version of AddmusicK).