WIP commit for generating hot patch VCMDs before $F4 $04 VCMDs
This currently accounts for VCMDs being placed before a note. In addition, the
presence of a $F1 VCMD immediately disqualifies this custom setup: this is
explicitly handling the zero EDL case where nothing else is accounted for.

This does not yet account for VCMDs being placed after an intro: for those
ones, the entire VCMD needs to be backed up and re-generated somewhere else.