Update AddmusicK Parser Version to 3 for this branch
This marks the beginning of the implementation for Codec's AddmusicK Beta. Note
that the upcoming code will be sourced, though also probably slightly modified,
from the C++ source, which is otherwise not public, unlike the rest of the
extant beta: I acquired the source from Codec themselves.

This branch will not be merged until all of the music changes are implemented.

This commit mentions #216.