Add the option to restart P-Switch and Star global songs on retrigger
This code was initially contributed by the JUMP team, with SimFan96 updating it
for AddmusicK 1.0.6. KevinM then made a version for 1.0.8, which Anas then
modified to be user definable as well as some adaptation for the modifications
that were made in the AMKFF fork, and I made some final adjustments. This
replicates vanilla SMW behavior where these songs restart when they are
retriggered.

This commit closes #193.