Optimize some zero-related utilizations in hardcoded 1DFA SFX
The first optimization stores a constant zero instead of using the accumulator
to get rid of a push/pop operation (storing to two direct page locations takes
the same number of bytes with the same constant, so we save two bytes). The
second optimization moves a zero storage to where the accumulator is used,
saving a byte in the process.