Replace two incs with an adc while setting noise frequencies
This operation normally would not be done because of a clrc operation. However,
within the context of the code here, the carry is always cleared before the
operation in question, saving a byte. This happens because the comparisons are
always intended to yield a result less than the intended value, which means the
carry is cleared. Note that I tend to be safety-oriented and therefore tend to
use clrc opcodes before an adc: this scenario is me realizing it is safe to not
do that. This kind of optimization is not always simple to do because the state
of the carry may not be known: in order for it to work, the clrc opcode must
service multiple branches and not be modified during the entire duration prior
to an adc or a sbc.