Try to build AddmusicK command line application with MinGW - Try 23
Initial cygwin test, no makefile parameters modified. I found a suspected later
MinGW-w64 build in there, which for me means it might be a better solution...
but the default may be just a cygwin case, so stay tuned. If it does build
successfully, though, then I'll accept it as a test case as long as it's not too
expensive in filesize.