Optimize EffectModifier to remove utilization of Y
Take advantage of a, (X) to save a byte by having !MusicPModChannels itself be
the value to use for the X register as well as utilizing direct DSP register
address references to remove a dependency on Y, saving a push/pop operation. In
addition, the direct page location has been changed to allow it to be used
slightly earlier when setting up an instrument to save a push/pop there too.