Relocate code to reduce the usage of jmp opcodes
Each removal of a jmp opcode will save at least a byte. Bonus points if there is
no preceding code to the pointer in question, which means instead of
transforming it into a bra and only saving a byte, two additional bytes are
saved from the completely removal of the jmp opcode.