Correct a fault in the adaptation of UseGainInstead
I realized that the accumulator did not contain the actual gain parameter, thus
I have to adapt the preceding RunRemoteCode code instead to send that over to
the target register earlier. As a trade off, the storages to $30+x are done in
one go with A and Y. The gain parameter being stored in Y as a side effect ends
up saving a byte.