Remove raw hex usage blockage for $FC VCMD
This was present for #amk 2 in order to prevent usage of it due to its purpose
having been changed from AMK beta from remote gain to remote code. All of the
other hex commands that used raw pointers were noticeably NOT blocked in this
kind of manner, so I feel this is fair game. I won't be doing the same for
anticipation gain's blockage removal, though.