Allow (!!) in parser versions 3 and up
(!!) was initially not going to be implemented in parser version 4 due to me not
wanting to grab anything that required a new hex command. However, due to the
way AMK Beta's Remote Gain was restored, this MML command can now make the cut
into parser version 3.

: and % are still cut from parser version 4, on that note: : because it requires
a new hex command, and % because of a conflict with binary notation in asar (I
have decided to reserve this for parsing a binary number).