Output a warning for using $FC with #amk 2 or #amk 3
Because this outputs an error on older versions of AddmusicK and the ability to
use it as a raw hex command was added on because of the restoration of remote
gain, I've turned this situation into a warning. I've also added a notice with
the warning that the command is five bytes, not two, because of the ability to
change the event type.