Use divideByTempoRatio in parser versions 4 and up
Turns out this call was effectively dummied out for prior parser versions by
immediately returning. If we are using parser version 4 or later, this should
only happen if the tempo ratio is 1, meaning dividing will effectively do
nothing (and also not force an error if fractional tick values are output on
the note lengths themselves).

Note that in order to properly do this, I had to add warnings to the note
length operations for when they output fractional values, because otherwise
errors would be output when using a fractional value under any circumstance.
These will be output under any parser version mostly so that if I do implement
 #105 for later parser versions, these warnings will indicate that they fall
under the conditions that #105 will attempt to rectify some time in the future,
though that will be a challenge in and of itself.

This commit closes #304.