Set !PSwitchStarRestart to default to !true
In some cases it's not practical to replicate the vanilla behavior, usually
because of modifications already made internally. On the other hand, this one
only affects the SNES side, and therefore can default to its vanilla behavior.
It also means that the user define itself needs to be noted that the default is
different for replicating the behavior on AddmusicK 1.0.8 and earlier.