VCMD -> Command terminology update for documentation
I'm going to let the VCMD (voice command) terminology stay for the code I added
in to the C++ side and the SPC side, particularly the latter since it existed in
the first place from a disassembly. However, to prevent user confusion for those
reading the documentation, which they are likely to be most commonly looking up,
I'm modifying some of the text to be more consistent with what is already said
in the documentation and doesn't use this kind of abbreviation.