Remove silent BRR blocks from the ends of non-looping default samples
Contributed by Pinci, with some extra editing by me. The BRR decoder outputs
samples so quiet at the end that the presence of these silent BRR blocks are
rendered redundant. Thus, we save $1B+$E1+$12+$2D+$12+$2D=$17A bytes by getting
rid of them.

Note that there has to be a zero or two at the very last nibble with a quiet
enough shift value on the header of the last BRR block: if this criteria is not
met, then a silent BRR block must be used. And if the decoder does not properly
silence after one BRR block, this silent block must be looped.