Add deep nesting to SPC side
This almost completely replaces the standard storage locations for $E6 and $E9
loops: only $C0+x is still used to store the nesting level. Instead, all nesting
data will be dynamically allocated when loading a song. For now, it's allocated
to $C000 because the dynamic part has not yet been implemented on both the C++
side and the SNES side (yes, really: the SNES side is getting involved because
it tracks where to load data on the SPC700).

Note that the loop type reads and writes are planned, but ultimately are not
implemented. This is because their detection is currently not needed unless loop
breaks are involved. I've still reserved the byte for now.

This commit mentions #302.