Optimize out the utilization of GetSampleTableAddress on SPC side
This optimization required modifications to the C++ side (to remove the setting
of !TabARAMRet), the SNES side (to replace the sample directory setup code) and
the SPC side (to remove the routine).

The sample directory setup can be simplified into storing the DSP register
writes into scratch RAM, then sending it as a two-byte block into $F2 on the SPC
side, which writes it directly to the DSP register specified. I've seen this
technique used before straight out of the IPL Boot ROM, and since this sound
driver uses a variant of that, it means I can use the same technique here.

This commit closes #326.