Have remote gain play nice with SFX
Currently there is no code to handle remote gain with SFX. AddmusicK Beta didn't
have this problem, but now that it is supported across all remote code types, it
can be a problem when interacting with the SFX. One of the memory locations has
been repurposed to be multi-bit in order to ensure that the restoration only
happens when it needs to.

This commit closes #331.