Add AddmusicM samples and sample groups
Four sample groups have been created, all of them prefixed with #AMMKisoPCM:
They replicate the default sets used for AddmusicM. #AMMKisoPCMS00 is the
default set. Note that the original SRCN IDs from Super Mario World are also
implemented here: this is because of how the global instruments are handled on
AddmusicK, as they are preconfigured to use the original IDs. Just make sure to
keep this in mind if you even remotely try to backport these...

KisoPCM refers to the original binary file that stored the samples. The core
BRR filenames match AddmusicK's because they had no individual filenames, but
the swappable BRR files keep their original flilenames.

06_empty.brr is omitted because it is an illegal BRR file, consisting of a
single $AA byte, not counting the loop point data.

All of them replace the #default and #optimized sample groups for AddmusicM
songs only.

This commit mentions #312. It does not close the issue due to a dependency on
 #31 in order to complete the sample pack restoration.