Redo the AddmusicM sample groups + update AddmusicM to AddmusicK notes
I decided to keep the sXX packs separate from the main #AMMKisoPCM sample
group, marking them as #AMMS01, #AMMS02 and #AMMS03 instead. #AMMS00 does
not exist because it has no new samples: it only has a single empty slot,
which was done because it could not be completely dummied out in the
original.

This commit closes #312.