Replace Bowser Defeated & Bowser Zoom Out original SMW ports
The reason why the replacement is being done now is due to the presence of a r39
and r18 in Bowser Defeated, which is triggering a fractional note length warning
in the default set. These are being replaced with new ports that, in addition to
keeping the original note data not counting error-related bugfixes (though with
some fine print), also replicates a bug in the original song where the pitch
fades were being interrupted by the phrase end markers. These are implemented by
simply commenting out one of the $DD VCMDs at the end: the reason they're
commented out, but not removed, is because they're part of the original note
data. The same case applies to rests because each phrase may have two rests in a
row, one from the end of a previous phrase and one from the beginning of the
other phrase: the original rest of the next phrase is commented out, and if
there is a leftover length, then a new rest is defined in its stead with the
adjusted length. Note that this kind of offset cannot be done with the original
phrase system: this is merely an unintentional side effect. There's also some
basic redundancy removals commented out accordingly. More optimizations could be
done, but for the time being I won't do so, mostly because they're not global
songs.

Bowser Zoom In was also analyzed and verified good by me (though it has some
unnecessary elements in it: specifically the $ED VCMDs).

This commit mentions #309.