Change negative note errors into a warning for #amk 3 and earlier
When I fixed the detection of the pitch, I accidentally broke some older ports
that would have technically worked due to a loophole in how a note byte was
detected. They have been allowed again, but with a warning, as the note would
otherwise likely not be audible under normal circumstances due to other problems
with the pitch calculation.

This commit closes #338.