Add a hot patch bit for skipping forced key offs on rests
This hot patch is only applied to Romi's Addmusic404, Addmusic405 and carol's
MORE.bin. On AddmusicM, it has an earlier version of a forced key off on rests.
What this does is that it causes key offs to only occur during readahead, which
is what vanilla SMW did. The placement is a little unusual because I want the
inRest variable to still do its work as needed as well as the SFX check.

As part of the addition, all presets are now two bytes, and an extra byte's
worth of writes were set up. The template has been moved down one byte.

This commit mentions #222 and closes #341.