Use chains of BRA opcodes instead in readahead whenever possible
Optimizing out JMP opcodes and chaining BRA opcodes will help reduce the
filesize. This is also because they're already used to go to a JMP opcode: if
the JMP opcode itself is in range, it can instead become a BRA chain, with the
BRA opcodes jumping a certain number of bytes at a time and linking with each
other wherever needed.