Use a mov1 opcode for handling Yoshi Drum APU1 commands
Since a TCLR/TSET opcode differs only by one bit, we can use mov1 opcodes
instead and let the carry determine what opcode it is. Thanks to the way the
checks work by ID, anything that is equal also sets the carry. Thus, we only
need to move back an additional byte to clear the carry and disable Yoshi Drums.