Turn the APU1/1DFA command set into a jump array
There are enough individual checks for IDs in order that it can be transformed
into an array on the fly. This saves a few bytes by doing so, but it's only
practical when SFX is enabled: the old check system is still used when !noSFX is
set to true because there are gaps in the IDs, and thus would actually yield a
larger size if the array was used.