Optimize ForceSFXEcho commands and use BRA instead of JMP
First, we use the carry to affect an ADC operation since we want a result of $FF
or zero (since there's 8-bit wraparound) to use for the first storage. Using
this instead of a BRA opcode saves a byte (we use one to set the carry in one
case, and the carry is cleared when jumping from the array by default).

Second, this command is no longer bound by a BRA opcode to the ForceSFXEcho
commands, so the JMP opcode becomes a BRA opcode and we save a second byte.